fix: corrige parsing de chart e sincronização de TravelTime
- ChartImporter: todos os TryParse usam CultureInfo.InvariantCulture
  (evita falha silenciosa com locale pt-BR onde '.' não é separador decimal)
- NoteData.Duration: double → float (consistente com Note.Duration)
  Remove cast (float) desnecessário em GameManager e GetSingle() em loaders
- ChartImporter: duração de hold não inclui offset duas vezes
  (TicksToSeconds(tick+sustain) - TicksToSeconds(tick), sem + offset)
- GameData.NoteSpeed: novo campo estático definido por GameManager._Ready
  garante que GameData.TravelTime == GameManager.TravelTime em qualquer configuração

+++ Scripts/LoadingScreen.cs
@@ -192,7 +192,7 @@ public partial class LoadingScreen : Control
 					if (n.TryGetProperty("time",     out var t))  nd.Time     = t.GetDouble();
 					if (n.TryGetProperty("lane",     out var l))  nd.Lane     = l.GetInt32();
 					if (n.TryGetProperty("isLong",   out var il)) nd.IsLong   = il.GetBoolean();
-					if (n.TryGetProperty("duration", out var d))  nd.Duration = d.GetDouble();
+					if (n.TryGetProperty("duration", out var d))  nd.Duration = d.GetSingle();
 					_chartNotes.Add(nd);
 				}
 				_chartNotes.Sort((a, b) => a.Time.CompareTo(b.Time));
